requirements.txt
Pillow>=9.4.0
python-Levenshtein>=0.20.9
vaik-text-recognition-trt-inference @ git+https://github.com/vaik-info/vaik-text-recognition-trt-inference.git@0e337183749e6a4bb12589436b2c0505faf8502a